Page 컴포넌트는 별도로 추출하도록 구성
