+++ app/(main)/groups/[groupId]/posts/(fullscreen)/create/page.tsx
@@ -1,18 +1,5 @@
-import { submitPost } from "@/features/post/api/createPost";
-import PostEditor from "@/features/post/components/PostEditor";
+import { PostCreatePage } from "@/features/post/screens/PostCreatePage";
 
-import TitleBackHeader from "@/shared/components/layout/headers/TitleBackHeader";
-
-async function PostCreatePage() {
-	return (
-		<div className="relative">
-			<TitleBackHeader title="내 물품 수정" />
-			<PostEditor
-				mode="create"
-				defaultValues={{ title: "", content: "", rentalFee: 0, feeUnit: "HOUR" }}
-				onSubmit={submitPost}
-			/>
-		</div>
-	);
+export default function Page() {
+	return <PostCreatePage />;
 }
-export default PostCreatePage;
